+++ angular-workspace/projects/ni/nimble-angular/table/nimble-table.directive.ts
@@ -1,10 +1,11 @@
 import { Directive, ElementRef, Input, OnDestroy, Renderer2 } from '@angular/core';
-import type { Table } from '@ni/nimble-components/dist/esm/table';
+import { type Table, tableTag } from '@ni/nimble-components/dist/esm/table';
 import type { TableRecord, TableFieldName, TableFieldValue, TableValidity, TableActionMenuToggleEventDetail, TableRowSelectionEventDetail, TableColumnConfigurationChangeEventDetail, TableColumnConfiguration } from '@ni/nimble-components/dist/esm/table/types';
 import { TableRowSelectionMode } from '@ni/nimble-components/dist/esm/table/types';
 import type { Observable, Subscription } from 'rxjs';
 
 export type { Table };
+export { tableTag };
 export type { TableActionMenuToggleEventDetail, TableRowSelectionEventDetail, TableColumnConfigurationChangeEventDetail, TableColumnConfiguration };
 export { TableRecord, TableFieldName, TableFieldValue, TableValidity, TableRowSelectionMode };
 
@@ -19,6 +20,15 @@ export class NimbleTableDirective<TData extends TableRecord = TableRecord> imple
         return this.dataObservable;
     }
 
+    /**
+     * An `Observable<TData[]>` that emits data that should be set on the table.
+     *
+     * When using `data$`, the timing of when the table has been updated with a new collection of data
+     * cannot be observed or await by an application. If an application needs to know when the data has
+     * been applied, explicitly call `setData()` rather than using the `data$` property. For example,
+     * calling `getSelectedRecordIds()` immediately after emitting a new value to the `data$` `Observable`
+     * may not take the new data into account when determining the selected records.
+     */
     @Input() public set data$(value: Observable<TData[]> | undefined) {
         this.dataSubscription?.unsubscribe();