+++ packages/nimble-components/src/table-column/text-base/group-header-view/index.ts
@@ -11,45 +11,29 @@ export abstract class TableColumnTextGroupHeaderViewBase<
     TColumnConfig = TableColumnWithPlaceholderColumnConfig
 > extends TableGroupHeaderView<TGroupValue, TColumnConfig> {
     /** @internal */
-    public textSpan!: HTMLElement;
-
-    /** @internal */
     @observable
     public hasOverflow = false;
 
     /**
-     * Returns the text to render in the cell when it contains a valid value (i.e. when shouldUsePlaceholder() is false).
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
+     * Text to render in the cell when it contains a valid value (i.e. when shouldUsePlaceholder is false).
      */
-    public abstract get text(): string;
+    @observable
+    public text = '';
 
     /**
-     * Returns the text to render in the cell when it contains an invalid value (i.e. when shouldUsePlaceholder() is true).
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
+     * Text to render in the cell when it contains an invalid value (i.e. when shouldUsePlaceholder is true).
      */
-    public abstract get placeholder(): string;
+    @observable
+    public placeholder = '';
 
     /**
      * Returns whether to display the placeholder value or the text value
-     * If the implementation has branching code paths then it must be marked with @volatile.
-     * https://www.fast.design/docs/fast-element/observables-and-state/#observable-features
-     * */
-    public abstract get shouldUsePlaceholder(): boolean;
+     */
+    @observable
+    public shouldUsePlaceholder = true;
 
     @volatile
     public get content(): string {
         return this.shouldUsePlaceholder ? this.placeholder : this.text;
-    }
-
-    /** @internal */
-    public updateTitleOverflow(): void {
-        this.hasOverflow = this.textSpan.offsetWidth < this.textSpan.scrollWidth;
-    }
-
-    /** @internal */
-    public clearTitleOverflow(): void {
-        this.hasOverflow = false;
     }
 }